Update VMwareHorizonClient.download.recipe
Happy Sunday!

Added a variable `HORIZON_MAJOR_VERSION` to determine the major version of the VMware Horizon Client downloaded as people may need specific major versions depending on their environments, so can override as necessary.

Have set this to `horizon_8` which is currently the highest major version.

This should not break compatibility with existing child recipes from yourself and @rtrouton